Split `Runtime` into `WorkerRuntime` and `Runtime`. Add emulator.
This commit splits the original `Runtime` implementation into two
distinct runtimes, where the functionality is specialized to the
particular use-case of the environment. This not only clarifies the
semantics of each, but also allows the orchestration `Runtime` to
start a `WorkerRuntime` instance without having to recurse on
itself as it would if they were the same type. This enables the
orchestration `Runtime` to trivially start an emulator if the
requisite configuration option is provided.

This commit also provides an implementation of an in-memory queue,
which can be used to emulate an asynchronous queue environment
in memory. This will enable tests for directives to be written
without having to worry about having a queue available in the
environment.